mms/install_genomes/requirements.txt
pip==21.3.1
biopython==1.81